Flyway Database Migrations added
Two migration files have been added.
V1 - to create initial tables
V2- add foreign keys

+++ backend_PharmaPulse/src/main/java/com/group03/backend_PharmaPulse/purchase/api/dto/PurchaseLineItemDTO.java
@@ -21,5 +21,5 @@ public class PurchaseLineItemDTO extends LineItemDTO {
     private LocalDate manufactureDate;
     private LocalDate  expiryDate;
     private BigDecimal unitPrice; //check and rename.Retail price of the product
-    private Integer purchaseInvoice;
+    private Long purchaseInvoice;
 }
